rpm tests
